backend/requirements.txt
Django>=5.0
djangorestframework
psycopg2-binary
gunicorn
python-dotenv
django-cors-headers
django-filter